+++ include/aquarius/client.hpp
@@ -1,8 +1,8 @@
 ﻿#pragma once
 #include <aquarius/core/asio.hpp>
 #include <aquarius/core/concepts.hpp>
-#include <aquarius/invoke/invoke_callback.hpp>
 #include <aquarius/core/logger.hpp>
+#include <aquarius/invoke/callback.hpp>
 #include <aquarius/response.hpp>
 #include <filesystem>
 #include <iostream>
