+++ include/aquarius/context/context.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <aquarius/context/impl/visitor.hpp>
 #include <aquarius/core/logger.hpp>
-#include <aquarius/session/xsession.hpp>
+#include <aquarius/session/session.hpp>
 #include <chrono>
 
 using namespace std::chrono_literals;
@@ -37,7 +37,7 @@ namespace aquarius
 		basic_context& operator=(const basic_context&) = delete;
 
 	public:
-		virtual error_code visit(impl::basic_message*, std::shared_ptr<xsession> session_ptr, error_code& ec) override
+		virtual error_code visit(impl::basic_message*, std::shared_ptr<basic_session> session_ptr, error_code& ec) override
 		{
 			XLOG_WARNING() << name_ << " maybe visit an unknown message!";
 
@@ -81,7 +81,7 @@ namespace aquarius
 
 		std::chrono::milliseconds timeout_;
 
-		std::shared_ptr<xsession> session_ptr_;
+		std::shared_ptr<basic_session> session_ptr_;
 	};
 
 } // namespace aquarius
