+++ include/aquarius/message/impl/visitable.hpp
@@ -10,7 +10,7 @@ namespace aquarius
 		class basic_message;
 	}
 
-	class xsession;
+	class basic_session;
 
 	class basic_context;
 } // namespace aquarius
@@ -27,12 +27,12 @@ namespace aquarius
 
 		public:
 			virtual error_code accept(flex_buffer_t& buffer, std::shared_ptr<basic_context> ctx,
-									  std::shared_ptr<xsession> session_ptr, error_code& ec) = 0;
+									  std::shared_ptr<basic_session> session_ptr, error_code& ec) = 0;
 		};
 
 		template <typename _Request>
 		static error_code accept_shared_impl(std::shared_ptr<_Request> req, std::shared_ptr<basic_context> ctx,
-											 std::shared_ptr<xsession> session_ptr, error_code& ec)
+											 std::shared_ptr<basic_session> session_ptr, error_code& ec)
 		{
 			using visitor_t = shared_visitor<_Request>;
 
@@ -46,7 +46,7 @@ namespace aquarius
 
 		template <typename _Request>
 		static error_code accept_bare_impl(_Request* req, std::shared_ptr<basic_context> ctx,
-										   std::shared_ptr<xsession> session_ptr, error_code& ec)
+										   std::shared_ptr<basic_session> session_ptr, error_code& ec)
 		{
 			using visitor_t = bare_visitor<aquarius::impl::basic_message>;
 
@@ -62,7 +62,7 @@ namespace aquarius
 
 #define DEFINE_VISITABLE_REQUEST()                                                                                     \
 	virtual aquarius::error_code accept(aquarius::flex_buffer_t& buffer, std::shared_ptr<aquarius::basic_context> ctx, \
-										std::shared_ptr<aquarius::xsession> session_ptr, aquarius::error_code& ec)     \
+										std::shared_ptr<aquarius::basic_session> session_ptr, aquarius::error_code& ec)     \
 	{                                                                                                                  \
 		this->from_binary(buffer, ec);                                                                                 \
 		if (ec)                                                                                                        \
@@ -72,7 +72,7 @@ namespace aquarius
 
 #define DEFINE_VISITABLE_RESPONSE()                                                                                    \
 	virtual aquarius::error_code accept(aquarius::flex_buffer_t& buffer, std::shared_ptr<aquarius::basic_context> ctx, \
-										std::shared_ptr<aquarius::xsession> session_ptr, aquarius::error_code& ec)     \
+										std::shared_ptr<aquarius::basic_session> session_ptr, aquarius::error_code& ec)     \
 	{                                                                                                                  \
 		this->from_binary(buffer, ec);                                                                                 \
 		if (ec)                                                                                                        \
@@ -83,7 +83,7 @@ namespace aquarius
 #define DEFINE_VISITABLE()                                                                                             \
 	virtual aquarius::error_code accept(                                                                               \
 		[[maybe_unused]] aquarius::flex_buffer_t& buffer, std::shared_ptr<aquarius::basic_context> ctx,                \
-		[[maybe_unused]] std::shared_ptr<aquarius::xsession> session_ptr, aquarius::error_code& ec)                    \
+		[[maybe_unused]] std::shared_ptr<aquarius::basic_session> session_ptr, aquarius::error_code& ec)                    \
 	{                                                                                                                  \
 		return accept_bare_impl(this, ctx, session_ptr, ec);                                                           \
 	}
