+++ test/unit_test.vcxproj
@@ -157,7 +157,7 @@
     <ClInclude Include="..\include\aquarius\client.hpp" />
     <ClInclude Include="..\include\aquarius\connect.hpp" />
     <ClInclude Include="..\include\aquarius\context.hpp" />
-    <ClInclude Include="..\include\aquarius\context\basic_context.hpp" />
+    <ClInclude Include="..\include\aquarius\context\context.hpp" />
     <ClInclude Include="..\include\aquarius\context\impl\visitor.hpp" />
     <ClInclude Include="..\include\aquarius\core\error_code.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\error_category.hpp" />
@@ -165,7 +165,7 @@
     <ClInclude Include="..\include\aquarius\core\impl\system_error_message.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\time_traits.hpp" />
     <ClInclude Include="..\include\aquarius\core\impl\type_traits.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\invoke_callback.hpp" />
+    <ClInclude Include="..\include\aquarius\invoke\callback.hpp" />
     <ClInclude Include="..\include\aquarius\message\header.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\custom.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\header.hpp" />
@@ -182,13 +182,13 @@
     <ClInclude Include="..\include\aquarius\router\message.hpp" />
     <ClInclude Include="..\include\aquarius\router\service.hpp" />
     <ClInclude Include="..\include\aquarius\router\session.hpp" />
-    <ClInclude Include="..\include\aquarius\session\xsession.hpp" />
+    <ClInclude Include="..\include\aquarius\session\session.hpp" />
     <ClInclude Include="..\include\aquarius\core\core.hpp" />
     <ClInclude Include="..\include\aquarius\core\elastic.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\invoke_context.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\invoke_message.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\invoke_service.hpp" />
-    <ClInclude Include="..\include\aquarius\invoke\invoke_session.hpp" />
+    <ClInclude Include="..\include\aquarius\invoke\context.hpp" />
+    <ClInclude Include="..\include\aquarius\invoke\message.hpp" />
+    <ClInclude Include="..\include\aquarius\invoke\service.hpp" />
+    <ClInclude Include="..\include\aquarius\invoke\session.hpp" />
     <ClInclude Include="..\include\aquarius\core\io_service_pool.hpp" />
     <ClInclude Include="..\include\aquarius\core\logger.hpp" />
     <ClInclude Include="..\include\aquarius\message\message.hpp" />
